DBZ-5359 Use kafka connect URI with default jolokia port
