+++ backend/src/main/java/io/debezium/configserver/rest/client/jolokia/JolokiaAttributes.java
@@ -3,7 +3,7 @@
  *
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.debezium.configserver.rest.client;
+package io.debezium.configserver.rest.client.jolokia;
 
 import java.io.IOException;
 import java.io.InputStream;
