+++ backend/src/main/resources/application.properties
@@ -10,9 +10,6 @@ quarkus.http.test-ssl-port=0
 # only enable the next line in order to allow CORS request from any domain for all modes (dev and prod, etc), e.g. for UI frontend development (see details in README.md), when you can't use the dev mode
 # quarkus.http.cors.origins=*
 
-# Jolokia endpoint inside Kafka Connect images
-jolokia.client.uri=http://dbzui-connect:8778/jolokia
-
 ## Test Env Configuration
 %test.quarkus.log.category."org.testcontainers".level=WARN
 %test.quarkus.log.category."com.github.dockerjava".level=WARN
